bots/telegram/webhooks/requirements.txt
absl-py==1.3.0
aio-pika==8.2.3
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiormq==6.4.2
aiosignal==1.3.1
anyio==4.3.0
APScheduler==3.9.1.post1
apturl==0.5.2
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.9.1
bidict==0.22.1
blinker==1.4
boto3==1.26.122
botocore==1.29.122
Brlapi==0.8.3
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
cloudpickle==2.2.1
colorama==0.4.4
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
command-not-found==0.3
confluent-kafka==1.9.2
constantly==23.10.4
cryptography==42.0.5
cupshelpers==1.0
cycler==0.11.0
daphne==4.1.0
dask==2022.10.2
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.2.12
dnspython==1.16.0
docopt==0.6.2
exceptiongroup==1.1.3
fbmessenger==6.0.0
fire==0.5.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
httpcore==1.0.4
httplib2==0.20.2
httptools==0.5.0
httpx==0.27.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==22.10.0
iniconfig==2.0.0
jeepney==0.7.1
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.1
jsonschema==4.17.3
keras==2.11.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.0
locket==1.0.0
louis==3.20.0
macaroonbakery==1.3.1
magic-filter==1.0.9
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
mattermostwrapper==2.2
more-itertools==8.10.0
msgpack==1.0.5
multidict==5.2.0
netifaces==0.11.0
networkx==2.6.3
numpy==1.24.3
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
packaging==20.9
pamqp==3.2.1
partd==1.4.0
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.0.0
portalocker==2.7.0
prompt-toolkit==3.0.28
protobuf==3.12.4
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.2
pydot==1.4.2
PyGObject==3.42.1
PyJWT==2.3.0
pykwalify==1.8.0
pymacaroons==0.13.0
pymongo==3.10.1
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.3
pytest==7.4.2
pytest-django==4.5.2
python-apt==2.4.0+ubuntu3
python-crfsuite==0.9.9
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-engineio==4.4.1
python-socketio==5.8.0
python-telegram-bot==21.0.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
questionary==1.10.0
randomname==0.1.5
rasa==3.5.6
rasa-sdk==3.5.1
redis==4.5.4
regex==2022.10.31
reportlab==3.6.8
requests==2.25.1
requests-mock==1.11.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rocketchat-API==1.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.8.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sentry-sdk==1.14.0
service-identity==24.1.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack-sdk==3.21.3
sniffio==1.3.1
SQLAlchemy==1.4.0
sqlparse==0.4.2
systemd-python==234
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-text==2.11.0
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
twilio==7.14.2
Twisted==24.3.0
txaio==23.1.1
typeguard==3.0.2
typing-utils==0.1.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.7.0
unattended-upgrades==0.1
urllib3==1.26.15
uvicorn==0.28.0
uvloop==0.17.0
wadllib==1.3.6
wcwidth==0.2.6
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==2.3.1
wrapt==1.15.0
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
zope.interface==6.2
